fx: move Flood
