+++ meson.build
@@ -98,6 +98,7 @@ sources = [
   'src/game/effect_routines/finish_level.c',
   'src/game/effect_routines/flicker.c',
   'src/game/effect_routines/flipmap.c',
+  'src/game/effect_routines/flood.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
   'src/game/fmv.c',
@@ -154,7 +155,6 @@ sources = [
   'src/game/objects/effects/bubble.c',
   'src/game/objects/effects/earthquake.c',
   'src/game/objects/effects/explosion.c',
-  'src/game/objects/effects/flood.c',
   'src/game/objects/effects/gunshot.c',
   'src/game/objects/effects/lara_effects.c',
   'src/game/objects/effects/missile.c',
