+++ src/game/effect_routines/flood.h
@@ -2,4 +2,4 @@
 
 #include "global/types.h"
 
-void Flood(ITEM_INFO *item);
+void FX_Flood(ITEM_INFO *item);
